requirements.txt
numpy==1.7.1
scipy==0.14.0
biopython==1.65
